backend/requirements.txt
flask
flask-cors
flask-jwt-extended
PyMySQL
bcrypt
pandas
numpy
scikit-learn
tensorflow
pillow
python-dotenv